Display network icon when confirmation transaction #4499
